requirements.txt
opencv-python
mediapipe
numpy
pandas
scikit-learn
scipy
matplotlib
tqdm
pyyaml
